loyalty-ai-backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
supabase>=2.4.0
sqlalchemy[asyncio]>=2.0.29
asyncpg>=0.29.0
langchain>=0.1.16
langgraph>=0.0.35
langchain-postgres>=0.0.4
openai>=1.20.0
instructor>=1.2.0
python-dotenv>=1.0.1
